Implemented respawning of all network-managed objects, even players
Work in progress, there's a new bug of spawning during re-spawn.
There's now a way better hierarchy using SpawnedObjects, which
contains children to be respawned, and it's used instead of prefabs,
so that it's easier to develop and modify the scene directly.
The Timer now properly invokes respawning. Player keeps their car.