Experimental Garage access and simple part/wheel detach syncing
The garage doesn't provide any persistence, but it can be tested.
The garage screen scaling is sub-optimal, hurting the back button.
Re-enabled the proper build scenes and added the Garage scene.
Disabled WebGL Auto Graphics API in order to fix issue of Linear.
Configured import settings of sqlite plugin within meta files.
The parts/wheels are now being synced via command when interpreted
to be required using physics at the client. This isn't a safe
solution, and the persistence of a detached component isn't solved.